+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## 0.5.0
+
+* Regenerated API for Chart.js 2.6.0
+
+* Support Dart 2.
+
+## 0.4.0+1
+
+* Allow the latest release of `package:func`.
+
 ## 0.4.0
 
 * Renamed `Chart` `toBase64` -> `toBase64Image`.